+++ controllers/common/php/log_mybooks.php
@@ -1,7 +1,5 @@
 <?php
 
-	//$_LOG['user_id'] = 199;
-
 	$_PAGE_TITLE = 'Mes achats';
 
 	$req = "SELECT `stock_id`, `stock_selling_date`,
@@ -18,7 +16,7 @@
 	
 	$stock = $_SQL->prepare($req);
 	$stock->bindValue('site_id',$_SITE['site_id'],PDO::PARAM_INT);
-	$stock->bindValue('user_id',$_LOG['user_id'],PDO::PARAM_INT);
+	$stock->bindValue('user_id',getLegacyVisitor()['user_id'],PDO::PARAM_INT);
 	$stock->execute() or error($stock->errorInfo());
 	while ($s = $stock->fetch(PDO::FETCH_ASSOC))
 	{
@@ -51,7 +49,7 @@
 		
 		<form action="/pages/export_to_csv" method="post">
 			<fieldset class="center">
-				<input type="hidden" name="filename" value="achats-'.$_SITE['site_name'].'-'.makeurl($_LOG['user_screen_name']).'">
+				<input type="hidden" name="filename" value="achats-'.$_SITE['site_name'].'-'.makeurl(getLegacyVisitor()['user_screen_name']).'">
 				<input type="hidden" name="header" value="'.htmlentities(json_encode($header)).'">
 				<input type="hidden" name="data" value="'.htmlentities(json_encode($export)).'">
 				<button type="submit">T&eacute;l&eacute;charger au format CSV</button>