+++ controllers/common/xhr/adm_article_files.php
@@ -7,7 +7,7 @@
 
     $controller = function() {
 
-        global $request, $_SQL, $_LOG, $site;
+        global $request, $_SQL, $site;
 
         $fm = new FileManager();
         $am = new ArticleManager();
@@ -34,7 +34,7 @@
 
                 $file_name = $request->request->get('name');
                 $file = $fm->create();
-                $fm->upload($file, $file_path, $file_name, $article->get('id'), $_LOG['user_id']);
+                $fm->upload($file, $file_path, $file_name, $article->get('id'), getLegacyVisitor()['user_id']);
 
                 // Return new table line
                 $r['success'] = "Le fichier &laquo;&nbsp;$file_name&nbsp;&raquo; a bien été associé à l'article &laquo;&nbsp;$article_title&nbsp;&raquo;.";
@@ -109,7 +109,7 @@
                 // Copy file into the files directory
                 try
                 {
-                    $fm->upload($file, $f['tmp_name'], $f['name'], $_POST['article_id'], $_LOG['user_id']);
+                    $fm->upload($file, $f['tmp_name'], $f['name'], $_POST['article_id'], getLegacyVisitor()['user_id']);
                     $file->markAsUpdated();
                 }
                 catch (Exception $e)